+++ CHANGELOG.md
@@ -13,6 +13,11 @@
   - `withReferenceScript` and `withStakingCredential` to add a reference script
     or staking credential to a payment
 - Export `Cooked.UtxoState`
+- A module `Cooked.ShowBS` to provide a Plutus-level analogue of `Show` with
+  `BuiltinString` as its codomain. This is very inefficient due to limitations
+  of `BuiltinString`, but potentially useful for "printf-debugging" of scripts.
+- An option `txOptEmulatorParamsModification` to temporarily change protocol
+  parameters for balancing and validation of a transaction
 
 ### Removed